Add update functionality to Product.
Splitting the ProductRequest classes into a Create and Update, as per your request on my question today. My only concern is the duplication of the validation rules, and just wondering your thoughts on the benefit of segregating vs centralizing for the validation logic.

Also update the postman with the update product PUT query.

Hope this helps.